Commit 3: code improvements.

+++ src/main/java/com/baeldung/carapi/controller/CarController.java
@@ -1,7 +1,6 @@
 package com.baeldung.carapi.controller;
 
 import com.baeldung.carapi.model.Car;
-import com.baeldung.carapi.repository.CarRepository;
 import com.baeldung.carapi.service.CarService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -19,6 +18,11 @@ public class CarController {
     @Autowired
     CarService carService;
 
+    /**
+     *
+     * @param ratingAbove
+     * @return returns all cars with rating above ratingAbove
+     */
     @GetMapping("/car")
     public ResponseEntity<List<Car>> getAllCars(
             @RequestParam(required = false) Optional<Integer> ratingAbove) {
@@ -32,6 +36,11 @@ public class CarController {
         }
     }
 
+    /**
+     *
+     * @param id
+     * @return
+     */
     @GetMapping("/car/{id}")
     public ResponseEntity<Car> getCarById(@PathVariable("id") int id) {
         Optional<Car> car = carService.getCarById(id);
@@ -42,6 +51,11 @@ public class CarController {
         }
     }
 
+    /**
+     *
+     * @param brand
+     * @return
+     */
     @GetMapping("/car/brand/{brand}")
     public ResponseEntity<List<Car>> findByBrand(@PathVariable("brand") String brand) {
         try {
@@ -54,6 +68,11 @@ public class CarController {
         }
     }
 
+    /**
+     *
+     * @param carReq
+     * @return
+     */
     @PostMapping("/car")
     public ResponseEntity<Car> createCar(@RequestBody Car carReq) {
         try {
@@ -64,6 +83,12 @@ public class CarController {
         }
     }
 
+    /**
+     *
+     * @param id
+     * @param carReq
+     * @return
+     */
     @PutMapping("/car/{id}")
     public ResponseEntity<Car> updateCar(@PathVariable("id") int id, @RequestBody Car carReq) {
         try {
@@ -71,20 +96,26 @@ public class CarController {
             if(car!=null)
                 return new ResponseEntity<>(car, HttpStatus.CREATED);
             else
-                return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
+                return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
         } catch (Exception e) {
             return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
+    /**
+     *
+     * @param id
+     * @return
+     */
     @DeleteMapping("/car/{id}")
     public ResponseEntity<HttpStatus> deleteCar(@PathVariable("id") int id) {
         try {
+            if (!carService.getCarById(id).isPresent())
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
             carService.deleteCar(id);
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         } catch (Exception e) {
             return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
-
 }
